3 errors with strategies

+++ VersionsStrategy.java
@@ -1,3 +1,4 @@
+import java.util.ArrayList;
 import java.util.List;
 import java.io.*;
 
@@ -17,7 +18,7 @@ class VolatileVersionsStrategy implements VersionsStrategy {
     private List<Document> dList;
 
     public VolatileVersionsStrategy() {
-
+        this.dList = new ArrayList<Document>();
     }
 
     public void putVersion(Document doc) {
@@ -25,7 +26,8 @@ class VolatileVersionsStrategy implements VersionsStrategy {
     }
 
     public Document getVersion() {
-        return this.dList.get(this.dList.size());
+
+        return this.dList.get(this.dList.size()-1);
     }
 
     public void setEntireHistory(List<Document> docList) {
@@ -37,7 +39,9 @@ class VolatileVersionsStrategy implements VersionsStrategy {
     }
 
     public void removeVersion() {
-        this.dList.remove(this.dList.size());
+        if(this.dList.size() > 1){
+            this.dList.remove(this.dList.size()-1);
+        }
     }
 }
 
@@ -45,7 +49,8 @@ class StableVersionsStrategy implements VersionsStrategy {
     private List<File> fileList;
     private List<Document> docList;
     public StableVersionsStrategy() {
-
+        this.fileList = new ArrayList<File>();
+        this.docList = new ArrayList<Document>();
     }
 
     public void putVersion(Document doc) {
@@ -64,26 +69,36 @@ class StableVersionsStrategy implements VersionsStrategy {
         }
     }
     public Document getVersion(){
-        return this.docList.get(this.docList.size());
+        if(this.docList.size() >= 1){
+            return this.docList.get(this.docList.size());
+        }
+        Document noDoc = new Document("", "", "", "", "");
+        return noDoc;
     }
     public void setEntireHistory(List<Document> dList){
-        for(int i=0;i<this.fileList.size();i++){
-            this.fileList.get(i).delete();
+        if(this.fileList != null){
+            for(int i=0;i<this.fileList.size();i++){
+                this.fileList.get(i).delete();
+            }
         }
         this.docList=dList;
-        this.fileList.clear();
-        for(int i=0;i<this.docList.size();i++){
-            String toString = this.docList.get(i).getContents();
-            File file;
-            FileWriter writer;
-            try {
-                file = new File("state"+i);
-                writer = new FileWriter(file);
-                writer.write(toString);
-                this.fileList.add(file);
-                writer.close();
-            } catch (IOException e) {
-                
+        if(this.fileList != null){
+            this.fileList.clear();
+        }
+        if(this.docList != null){
+            for(int i=0;i<this.docList.size();i++){
+                String toString = this.docList.get(i).getContents();
+                File file;
+                FileWriter writer;
+                try {
+                    file = new File("state"+i);
+                    writer = new FileWriter(file);
+                    writer.write(toString);
+                    this.fileList.add(file);
+                    writer.close();
+                } catch (IOException e) {
+                    
+                }
             }
         }
     }
@@ -91,8 +106,12 @@ class StableVersionsStrategy implements VersionsStrategy {
         return this.docList;
     }
     public void removeVersion(){
-        this.fileList.get(this.fileList.size()).delete();
-        this.fileList.remove(this.fileList.size());
-        this.docList.remove(this.docList.size());
+        if(this.fileList != null){
+            if(this.fileList.size() > 1){
+                this.fileList.get(this.fileList.size()).delete();
+                this.fileList.remove(this.fileList.size());
+                this.docList.remove(this.docList.size());
+            }
+        }
     }
 }
